# Optimize schedule.md