--- a/kevm-pyk/src/kevm_pyk/cli.py
+++ b/kevm-pyk/src/kevm_pyk/cli.py
@@ -300,7 +300,6 @@ class KEVMCLIArgs(KCLIArgs):
     @cached_property
     def evm_chain_args(self) -> ArgumentParser:
         schedules = (
-            'DEFAULT',
             'FRONTIER',
             'HOMESTEAD',
             'TANGERINE_WHISTLE',