--- a/kevm-pyk/src/tests/integration/test_conformance.py
+++ b/kevm-pyk/src/tests/integration/test_conformance.py
@@ -77,7 +77,7 @@ REST_VM_TESTS: Final = tuple(test_file for test_file in ALL_VM_TESTS if test_fil
     ids=[str(test_file.relative_to(VM_TEST_DIR)) for test_file in VM_TESTS],
 )
 def test_vm(test_file: Path) -> None:
-    _test(test_file, 'DEFAULT', 'VMTESTS', 1, True)
+    _test(test_file, 'HOMESTEAD', 'VMTESTS', 1, True)
 
 
 @pytest.mark.skip(reason='failing / slow VM tests')
@@ -87,7 +87,7 @@ def test_vm(test_file: Path) -> None:
     ids=[str(test_file.relative_to(VM_TEST_DIR)) for test_file in REST_VM_TESTS],
 )
 def test_rest_vm(test_file: Path) -> None:
-    _test(test_file, 'DEFAULT', 'VMTESTS', 1, True)
+    _test(test_file, 'HOMESTEAD', 'VMTESTS', 1, True)
 
 
 BCHAIN_NEW_TEST_DIR: Final = TEST_DIR / 'BlockchainTests/GeneralStateTests'