--- a/package/test-package.sh
+++ b/package/test-package.sh
@@ -7,16 +7,16 @@ kevm --help
 kevm version
 
 kevm run tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmArithmeticTest/add0.json --target llvm \
-    --mode VMTESTS --schedule DEFAULT --chainid 1                                                         \
+    --mode VMTESTS --schedule HOMESTEAD --chainid 1                                                       \
     > tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmArithmeticTest/add0.json.llvm-out         \
     || git --no-pager diff --no-index --ignore-all-space -R tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmArithmeticTest/add0.json.llvm-out tests/templates/output-success-llvm.json
 rm -rf tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmArithmeticTest/add0.json.llvm-out
 
 kevm run tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmSystemOperations/TestNameRegistrator.json \
-    --target llvm --mode VMTESTS --schedule DEFAULT --chainid 1
+    --target llvm --mode VMTESTS --schedule HOMESTEAD --chainid 1
 
 kevm run tests/ethereum-tests/LegacyTests/Constantinople/VMTests/vmSystemOperations/TestNameRegistrator.json \
-    --target haskell-standalone --mode VMTESTS --schedule DEFAULT --chainid 1
+    --target haskell-standalone --mode VMTESTS --schedule HOMESTEAD --chainid 1
 
 kevm kast tests/interactive/log3_MaxTopic_d0g0v0.json --target llvm > tests/interactive/log3_MaxTopic_d0g0v0.json.parse-out
 git --no-pager diff --no-index --ignore-all-space -R tests/interactive/log3_MaxTopic_d0g0v0.json.parse-out tests/interactive/log3_MaxTopic_d0g0v0.json.parse-expected