--- a/tests/specs/benchmarks/ecrecover00-siginvalid-spec.k
+++ b/tests/specs/benchmarks/ecrecover00-siginvalid-spec.k
@@ -9,6 +9,7 @@ module ECRECOVER00-SIGINVALID-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>