--- a/tests/specs/benchmarks/ecrecover00-sigvalid-spec.k
+++ b/tests/specs/benchmarks/ecrecover00-sigvalid-spec.k
@@ -9,6 +9,7 @@ module ECRECOVER00-SIGVALID-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>