--- a/tests/specs/benchmarks/ecrecoverloop00-sigs-valid-spec.k
+++ b/tests/specs/benchmarks/ecrecoverloop00-sigs-valid-spec.k
@@ -9,6 +9,7 @@ module ECRECOVERLOOP00-SIGS-VALID-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>