--- a/tests/specs/benchmarks/ecrecoverloop02-sig0-invalid-spec.k
+++ b/tests/specs/benchmarks/ecrecoverloop02-sig0-invalid-spec.k
@@ -9,6 +9,7 @@ module ECRECOVERLOOP02-SIG0-INVALID-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>