--- a/tests/specs/benchmarks/encode-keccak00-spec.k
+++ b/tests/specs/benchmarks/encode-keccak00-spec.k
@@ -9,6 +9,7 @@ module ENCODE-KECCAK00-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>