--- a/tests/specs/benchmarks/requires01-a0le0-spec.k
+++ b/tests/specs/benchmarks/requires01-a0le0-spec.k
@@ -9,6 +9,7 @@ module REQUIRES01-A0LE0-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>