--- a/tests/specs/benchmarks/staticloop00-a0lt10-spec.k
+++ b/tests/specs/benchmarks/staticloop00-a0lt10-spec.k
@@ -9,6 +9,7 @@ module STATICLOOP00-A0LT10-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>