--- a/tests/specs/benchmarks/storagevar02-overflow-spec.k
+++ b/tests/specs/benchmarks/storagevar02-overflow-spec.k
@@ -9,6 +9,7 @@ module STORAGEVAR02-OVERFLOW-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>