--- a/tests/specs/benchmarks/storagevar03-spec.k
+++ b/tests/specs/benchmarks/storagevar03-spec.k
@@ -9,6 +9,7 @@ module STORAGEVAR03-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>