--- a/tests/specs/benchmarks/structarg00-spec.k
+++ b/tests/specs/benchmarks/structarg00-spec.k
@@ -9,6 +9,7 @@ module STRUCTARG00-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
     <ethereum>
       <evm>