--- a/tests/specs/bihu/forwardToHotWallet-failure-1-spec.k
+++ b/tests/specs/bihu/forwardToHotWallet-failure-1-spec.k
@@ -10,6 +10,7 @@ module FORWARDTOHOTWALLET-FAILURE-1-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>