--- a/tests/specs/bihu/forwardToHotWallet-failure-3-spec.k
+++ b/tests/specs/bihu/forwardToHotWallet-failure-3-spec.k
@@ -10,6 +10,7 @@ module FORWARDTOHOTWALLET-FAILURE-3-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>