--- a/tests/specs/bihu/forwardToHotWallet-success-2-spec.k
+++ b/tests/specs/bihu/forwardToHotWallet-success-2-spec.k
@@ -10,6 +10,7 @@ module FORWARDTOHOTWALLET-SUCCESS-2-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> BYZANTIUM </schedule>
+    <scheduleTuple> getSchedule(BYZANTIUM) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>