--- a/tests/specs/erc20/ds/allowance-spec.k
+++ b/tests/specs/erc20/ds/allowance-spec.k
@@ -10,6 +10,7 @@ module ALLOWANCE-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>