--- a/tests/specs/erc20/ds/approve-failure-spec.k
+++ b/tests/specs/erc20/ds/approve-failure-spec.k
@@ -10,6 +10,7 @@ module APPROVE-FAILURE-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>