--- a/tests/specs/erc20/ds/approve-success-spec.k
+++ b/tests/specs/erc20/ds/approve-success-spec.k
@@ -10,6 +10,7 @@ module APPROVE-SUCCESS-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> CONSTANTINOPLE </schedule>
+    <scheduleTuple> getSchedule(CONSTANTINOPLE) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>