--- a/tests/specs/erc20/ds/balanceOf-spec.k
+++ b/tests/specs/erc20/ds/balanceOf-spec.k
@@ -10,6 +10,7 @@ module BALANCEOF-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>