--- a/tests/specs/erc20/ds/transfer-failure-1-b-spec.k
+++ b/tests/specs/erc20/ds/transfer-failure-1-b-spec.k
@@ -10,6 +10,7 @@ module TRANSFER-FAILURE-1-B-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> CONSTANTINOPLE </schedule>
+    <scheduleTuple> getSchedule(CONSTANTINOPLE) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>