--- a/tests/specs/erc20/ds/transferFrom-failure-2-b-spec.k
+++ b/tests/specs/erc20/ds/transferFrom-failure-2-b-spec.k
@@ -10,6 +10,7 @@ module TRANSFERFROM-FAILURE-2-B-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> ISTANBUL </schedule>
+    <scheduleTuple> getSchedule(ISTANBUL) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>