--- a/tests/specs/erc20/ds/transferFrom-success-1-spec.k
+++ b/tests/specs/erc20/ds/transferFrom-success-1-spec.k
@@ -10,6 +10,7 @@ module TRANSFERFROM-SUCCESS-1-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> CONSTANTINOPLE </schedule>
+    <scheduleTuple> getSchedule(CONSTANTINOPLE) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>