--- a/tests/specs/erc20/hkg/approve-spec.k
+++ b/tests/specs/erc20/hkg/approve-spec.k
@@ -10,6 +10,7 @@ module APPROVE-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> CONSTANTINOPLE </schedule>
+    <scheduleTuple> getSchedule(CONSTANTINOPLE) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>