--- a/tests/specs/erc20/hkg/transfer-success-2-spec.k
+++ b/tests/specs/erc20/hkg/transfer-success-2-spec.k
@@ -10,6 +10,7 @@ module TRANSFER-SUCCESS-2-SPEC
     <exit-code> 1 </exit-code>
     <mode> NORMAL </mode>
     <schedule> CONSTANTINOPLE </schedule>
+    <scheduleTuple> getSchedule(CONSTANTINOPLE) </scheduleTuple>
     <useGas> true </useGas>
 
     <ethereum>