… .github/workflows/test-pr.yml
… README.md
… kevm-pyk/pyproject.toml
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/driver.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/evm.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/gas.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/optimizations.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/schedule.md
… kevm-pyk/src/kevm_pyk/kproj/evm-semantics/state-utils.md
… package/version
… tests/specs/erc20/hkg/transferFrom-success-2-spec.k
… tests/specs/examples/erc20-spec.md
… tests/specs/examples/erc721-spec.md
… tests/specs/examples/solidity-code-spec.md
… tests/specs/examples/storage-spec.md
… tests/specs/examples/sum-to-n-spec.k
… tests/specs/functional/infinite-gas-spec.k
… tests/specs/functional/lemmas-spec.k
… tests/specs/kontrol/test-allowchangestest-testallow-0-spec.k
… tests/specs/kontrol/test-allowchangestest-testallow_fail-0-spec.k
… tests/specs/kontrol/test-allowchangestest-testfailallowcallstoaddress-0-spec.k
… tests/specs/kontrol/test-allowchangestest-testfailallowchangestostorage-0-spec.k
… tests/specs/kontrol/test-arithmetictest-test_max1_broken-uint256-uint256-0-spec.k
… tests/specs/kontrol/test-arithmetictest-test_wdiv_rounding-uint256-uint256-0-spec.k
… tests/specs/kontrol/test-arithmetictest-test_wmul_wdiv_inverse_underflow-uint256-uint256-0-spec.k
… tests/specs/kontrol/test-arithmetictest-test_wmul_weakly_increasing_positive-uint256-uint256-0-spec.k
… tests/specs/kontrol/test-countertest-testincrement-0-spec.k
… tests/specs/kontrol/test-emitcontracttest-testexpectemit-0-spec.k
… tests/specs/kontrol/test-emitcontracttest-testexpectemitcheckemitter-0-spec.k
… tests/specs/kontrol/test-emitcontracttest-testexpectemitdonotcheckdata-0-spec.k