This is a remake of #2335. That PR optimized the schedule lookup functions in schedule.md and deleted the DEFAULT schedule.

This PR takes a slightly different approach to get the same goal. We create a new sort, ScheduleTuple, which contains all of the values of all the schedule constants and schedule flags for a given schedule. We create an operator which computes a ScheduleTuple for a given Schedule via the existing code in schedule.md. Then we create getters for each individual member of the tuple (implicitly). Code that needs access to schedule data gets it by means of calling the getters on a value of sort `ScheduleTuple` from throughout the semantics, thus considerably simplifying the hot path while maintaining the same code to compute the actual values for the schedule.

Performance data:

On the master branch, we spend 35m21s executing the GeneralStateTests, of which 13% is spent in `_<_>`. On this branch, we spend 26m46s executing the GeneralStateTests, of which 2.5% is spent in ScheduleTuple getters.