@ehildenb 2024-03-28
@dwightguth please post performance information about this vs master if it's supposed to be an optimization.