@ehildenb 2024-04-02
Yes please, we sohuld check that it does not cause a regression in symbolic execution performance, using the same tables/techniques taht the HBB team uses.