@dwightguth 2024-04-05
| Test                                                                                      | schedule3 time    | master time       | (schedule3/master) time 
|-------------------------------------------------------------------------------------------|-------------------|-------------------|-------------------------
| erc20/ds/transferFrom-failure-1-d-spec.k                                                  | 84.85             | 89.64             | 0.9465640339134315      
| erc20/ds/transferFrom-failure-1-b-spec.k                                                  | 184.32            | 191.11            | 0.9644707236669979      
| mcd/cat-exhaustiveness-spec.k                                                             | 158.8             | 153.42            | 1.0350671359666277      
| erc20/ds/transfer-success-2-spec.k                                                        | 84.43             | 81.51             | 1.0358238252975096      
| mcd/dsvalue-peek-pass-rough-spec.k                                                        | 70.24             | 67.81             | 1.0358354225040554      
| bihu/forwardToHotWallet-success-2-spec.k                                                  | 65.59             | 63.23             | 1.037324055037166       
| benchmarks/dynamicarray00-spec.k                                                          | 55.28             | 53.28             | 1.0375375375375375      
| kontrol/test-emitcontracttest-testexpectemitdonotcheckdata-0-spec.k                       | 62.21             | 59.95             | 1.037698081734779       
| erc20/ds/balanceOf-spec.k                                                                 | 57.71             | 55.61             | 1.0377629922675777      
| bihu/forwardToHotWallet-failure-4-spec.k                                                  | 132.72            | 127.83            | 1.0382539310021122      
| erc20/hkg/approve-spec.k                                                                  | 55.62             | 53.52             | 1.0392376681614348      
| benchmarks/ecrecover00-sigvalid-spec.k                                                    | 81.98             | 78.87             | 1.0394319766704705      
| benchmarks/storagevar02-overflow-spec.k                                                   | 63.13             | 60.72             | 1.0396903820816865      
| mcd/flipper-ttl-pass-spec.k                                                               | 73.3              | 70.43             | 1.0407496805338632      
| erc20/ds/transfer-success-1-spec.k                                                        | 91.64             | 87.98             | 1.041600363719027       
| mcd/vat-addui-pass-spec.k                                                                 | 178.47            | 171.33            | 1.0416739625284537      
| benchmarks/address00-spec.k                                                               | 51.97             | 49.88             | 1.0419005613472332      
| mcd/vat-subui-fail-rough-spec.k                                                           | 95.48             | 91.63             | 1.0420168067226891      
| benchmarks/keccak00-spec.k                                                                | 59.62             | 57.19             | 1.0424899457947194      
| kontrol/test-emitcontracttest-testexpectemitcheckemitter-0-spec.k                         | 61.98             | 59.43             | 1.0429076224129228      
| erc20/ds/allowance-spec.k                                                                 | 62.33             | 59.76             | 1.0430053547523428      
| kontrol/test-allowchangestest-testallow-0-spec.k                                          | 64.16             | 61.46             | 1.0439310120403513      
| erc20/hkg/totalSupply-spec.k                                                              | 49.45             | 47.36             | 1.0441300675675675      
| benchmarks/storagevar02-nooverflow-spec.k                                                 | 63.9              | 61.18             | 1.0444589735207583      
| kontrol/test-expectcalltest-testexpectregularcall-0-spec.k                                | 62.21             | 59.56             | 1.0444929482874412      
| kontrol/test-arithmetictest-test_wdiv_rounding-uint256-uint256-0-spec.k                   | 73.6              | 70.45             | 1.0447125621007807      
| mcd/dstoken-approve-fail-rough-spec.k                                                     | 129.91            | 124.33            | 1.0448805598005309      
| mcd/vat-subui-pass-rough-spec.k                                                           | 172.57            | 165.15            | 1.0449288525582803      
| mcd/dai-symbol-pass-spec.k                                                                | 66.37             | 63.51             | 1.0450322783813573      
| kontrol/test-expectreverttest-test_expectrevert_message-0-spec.k                          | 77.21             | 73.87             | 1.0452145661296872      
| kontrol/test-expectreverttest-test_expectrevert_bytes4-0-spec.k                           | 163.84            | 156.67            | 1.0457649837237506      
| benchmarks/ecrecover00-siginvalid-spec.k                                                  | 78.03             | 74.61             | 1.04583835946924        
| kontrol/test-expectcalltest-testexpectstaticcall-0-spec.k                                 | 62.21             | 59.47             | 1.0460736505801245      
| benchmarks/requires01-a0le0-spec.k                                                        | 112.49            | 107.47            | 1.0467107099655717      
| mcd/dsvalue-read-pass-summarize-spec.k                                                    | 65.84             | 62.9              | 1.0467408585055644      
| benchmarks/storagevar03-spec.k                                                            | 56.19             | 53.65             | 1.0473438956197576      
| erc20/hkg/transfer-failure-1-spec.k                                                       | 74.85             | 71.44             | 1.0477323628219484      
| erc20/hkg/allowance-spec.k                                                                | 54.28             | 51.79             | 1.0480787796871984      
| mcd/dsvalue-read-pass-spec.k                                                              | 73.14             | 69.76             | 1.0484518348623852      
| benchmarks/bytes00-spec.k                                                                 | 63.4              | 60.46             | 1.048627191531591       
| kontrol/test-allowchangestest-testfailallowcallstoaddress-0-spec.k                        | 64.29             | 61.27             | 1.0492900277460422      
| kontrol/test-expectreverttest-testfail_expectrevert_false-0-spec.k                        | 106.94            | 101.91            | 1.0493572760278678      
| kontrol/test-storetest-teststoreload-0-spec.k                                             | 64.14             | 61.07             | 1.0502701817586377      
| examples/erc20-spec.md                                                                    | 301.55            | 287.0             | 1.0506968641114982      
| bihu/forwardToHotWallet-failure-3-spec.k                                                  | 100.14            | 95.13             | 1.0526647745190791      
| benchmarks/structarg01-spec.k                                                             | 70.03             | 66.51             | 1.0529243722748458      
| kontrol/test-arithmetictest-test_max1_broken-uint256-uint256-0-spec.k                     | 59.46             | 56.42             | 1.053881602268699       
| examples/solidity-code-spec.md                                                            | 218.88            | 207.6             | 1.0543352601156069      
| kontrol/test-allowchangestest-testfailallowchangestostorage-0-spec.k                      | 96.2              | 90.94             | 1.0578403342863427      
| mcd/vat-deny-diff-fail-rough-spec.k                                                       | 127.16            | 120.11            | 1.0586961951544418      
| kontrol/test-storetest-testaccesses-0-spec.k                                              | 64.67             | 61.0              | 1.0601639344262295      
| erc20/hkg/transfer-success-2-spec.k                                                       | 62.25             | 58.67             | 1.061019260269303       
| benchmarks/storagevar00-spec.k                                                            | 54.16             | 50.97             | 1.0625858348047872      
| benchmarks/encode-keccak00-spec.k                                                         | 62.58             | 58.89             | 1.0626591951095261      
| kontrol/test-expectreverttest-testfail_expectrevert_failandsuccess-0-spec.k               | 76.15             | 71.59             | 1.0636960469339294      
| kontrol/test-allowchangestest-testallow_fail-0-spec.k                                     | 64.7              | 60.81             | 1.06396974181878        
| mcd/cat-file-addr-pass-rough-spec.k                                                       | 71.68             | 67.27             | 1.065556711758585       
| benchmarks/structarg00-spec.k                                                             | 61.68             | 57.84             | 1.066390041493776       
| erc20/hkg/transfer-failure-2-spec.k                                                       | 72.53             | 67.91             | 1.0680312177882492      
| kontrol/test-arithmetictest-test_wmul_weakly_increasing_positive-uint256-uint256-0-spec.k | 96.7              | 90.5              | 1.0685082872928178      
| benchmarks/storagevar01-spec.k                                                            | 60.2              | 56.34             | 1.068512602058928       
| benchmarks/requires01-a0gt0-spec.k                                                        | 55.99             | 52.05             | 1.0756964457252642      
| kontrol/test-owneruponlytest-testincrementasowner-0-spec.k                                | 99.86             | 91.96             | 1.0859069160504569      
| kontrol/test-countertest-testincrement-0-spec.k                                           | 129.25            | 118.67            | 1.0891547990224995      
| TOTAL                                                                                     | 5636.509999999999 | 5391.650000000002 | 1.0454146689788835